backend/requirements.txt
openseespy>=3.4.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
numpy>=1.21.0
python-dotenv>=1.0.0
httpx>=0.24.0
python-multipart>=0.0.5
reportlab>=4.0.0
